Add building id constants
